+++ src/pages/index.jsx
@@ -1,22 +1,10 @@
-import React, { useEffect } from 'react'
-import { useRouter } from 'next/router'
+import React from 'react'
 import Layout from '../components/Layout'
-import { localeList, defaultLang } from '../translations'
+import { useLangageRedirect } from '../utils'
 
 const IndexPage = () => {
-  const router = useRouter()
-  useEffect(() => {
-    const acceptLangs = [navigator.language, ...navigator.languages].map((l) => l.split('-')[0])
-    let langToNavigate = defaultLang
-    for (const prefferedLang of acceptLangs) {
-      if (localeList.includes(prefferedLang)) {
-        langToNavigate = prefferedLang
-        break
-      }
-    }
-    router.replace('/[lang]/', `/${langToNavigate}/`)
-  }, [])
-  return <Layout></Layout>
+  useLangageRedirect('/')
+  return <Layout />
 }
 
 export default IndexPage
